Add files via upload
The above analysis could be perform on different locations and on different product types depending on the choice of features or what is desired by the analyst. The function i used above could be readjusted and use again. These functions made it easier to analyse any product in relation to other features.